@@ -1,7 +1,7 @@
 package dev.kizuna.api.events.impl;
 
 import dev.kizuna.api.events.Event;
-import net.minecraft.client.render.BlockBreakingInfo;
+import net.minecraft.entity.player.BlockBreakingInfo;
 import net.minecraft.util.math.BlockPos;
 
 public class WorldBreakEvent extends Event {
