@@ -10,6 +10,7 @@ import dev.kizuna.mod.modules.impl.render.NoRender;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.hud.InGameHud;
+import net.minecraft.client.render.RenderTickCounter;
 import net.minecraft.util.math.Vec3d;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -34,12 +35,12 @@ public class MixinInGameHud {
 	}
 
 	@Inject(method = "renderStatusEffectOverlay", at = @At("HEAD"), cancellable = true)
-	private void onRenderStatusEffectOverlay(DrawContext context, CallbackInfo ci) {
+	private void onRenderStatusEffectOverlay(DrawContext context, RenderTickCounter tickCounter, CallbackInfo ci) {
 		if (NoRender.INSTANCE.isOn() && NoRender.INSTANCE.potionsIcon.getValue()) ci.cancel();
 	}
 
-	@Inject(at = {@At(value = "INVOKE", target = "Lcom/mojang/blaze3d/systems/RenderSystem;enableBlend()V", remap = false, ordinal = 3)}, method = {"render(Lnet/minecraft/client/gui/DrawContext;F)V"})
-	private void onRender(DrawContext context, float tickDelta, CallbackInfo ci) {
+	@Inject(method = "render", at = @At("HEAD"))
+	private void onRender(DrawContext context, RenderTickCounter tickCounter, CallbackInfo ci) {
 		Kawaii.MODULE.render2D(context);
 	}
 
@@ -50,10 +51,10 @@ public class MixinInGameHud {
 		}
 	}
 
-    @ModifyArg(method="renderHotbar",at=@At(value="INVOKE",target="Lnet/minecraft/client/gui/DrawContext;drawGuiTexture(Lnet/minecraft/util/Identifier;IIII)V",ordinal = 1),index = 1)
+	@ModifyArg(method="renderHotbar",at=@At(value="INVOKE",target="Lnet/minecraft/client/gui/DrawContext;drawGuiTexture(Lnet/minecraft/util/Identifier;IIII)V",ordinal = 1),index = 1)
 	private int selectedSlotX(int x){
 		if (ClientSetting.INSTANCE.hotbar()) {
-            double hotbarX = ClientSetting.animation.get(x, ClientSetting.INSTANCE.hotbarTime.getValueInt(), ClientSetting.INSTANCE.animEase.getValue());
+			double hotbarX = ClientSetting.animation.get(x, ClientSetting.INSTANCE.hotbarTime.getValueInt(), ClientSetting.INSTANCE.animEase.getValue());
 			return (int) hotbarX;
 		}
 		return(x);
@@ -68,7 +69,7 @@ public class MixinInGameHud {
 	private double offsetCrosshairY;
 
 	@Inject(method = "renderCrosshair", at = @At("HEAD"), cancellable = true)
-	private void onRenderCrosshairBegin(DrawContext context, CallbackInfo ci) {
+	private void onRenderCrosshairBegin(DrawContext context, RenderTickCounter tickCounter, CallbackInfo ci) {
 		if (Crosshair.INSTANCE.isOn()) {
 			Crosshair.INSTANCE.draw(context);
 			ci.cancel();
