@@ -2,10 +2,10 @@ package dev.kizuna.asm.mixins;
 
 import dev.kizuna.mod.modules.impl.misc.Tips;
 import net.minecraft.block.ShulkerBoxBlock;
-import net.minecraft.client.item.TooltipContext;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.tooltip.TooltipType;
 import net.minecraft.text.Text;
-import net.minecraft.world.BlockView;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -16,7 +16,7 @@ import java.util.List;
 @Mixin(ShulkerBoxBlock.class)
 public class MixinShulkerBoxBlock {
     @Inject(method = "appendTooltip", at = @At("HEAD"), cancellable = true)
-    private void onAppendTooltip(ItemStack stack, BlockView view, List<Text> tooltip, TooltipContext options, CallbackInfo info) {
-        if (Tips.INSTANCE.isOn() && Tips.INSTANCE.shulkerViewer.getValue()) info.cancel();
+    private void onAppendTooltip(ItemStack stack, Item.TooltipContext context, List<Text> tooltip, TooltipType options, CallbackInfo ci) {
+        if (Tips.INSTANCE.isOn() && Tips.INSTANCE.shulkerViewer.getValue()) ci.cancel();
     }
 }
