@@ -31,9 +31,7 @@ public class ClickTP extends Module {
             if (hitResult.getType() == HitResult.Type.BLOCK) {
                 BlockPos pos = ((BlockHitResult) hitResult).getBlockPos();
                 Direction side = ((BlockHitResult) hitResult).getSide();
-                Hand hand = mc.player.getStackInHand(Hand.MAIN_HAND).getItem().isSuitableFor(mc.world.getBlockState(pos)) ? Hand.MAIN_HAND : Hand.OFF_HAND;
-
-                if (mc.world.getBlockState(pos).onUse(mc.world, mc.player, hand, (BlockHitResult) hitResult) != ActionResult.PASS) return;
+                if (mc.world.getBlockState(pos).onUse(mc.world, mc.player, (BlockHitResult) hitResult) != ActionResult.PASS) return;
 
                 BlockState state = mc.world.getBlockState(pos);
 
