@@ -86,16 +86,16 @@ public class PearlPhase extends Module {
 		AutoPearl.throwing = true;
 		int pearl;
 		if (mc.player.getMainHandStack().getItem() == Items.ENDER_PEARL) {
-			sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id));
+			sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id, mc.player.getYaw(), mc.player.getPitch()));
 		} else if (AutoPearl.INSTANCE.inventory.getValue() && (pearl = InventoryUtil.findItemInventorySlot(Items.ENDER_PEARL)) != -1) {
 			InventoryUtil.inventorySwap(pearl, mc.player.getInventory().selectedSlot);
-			sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id));
+			sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id, mc.player.getYaw(), mc.player.getPitch()));
 			InventoryUtil.inventorySwap(pearl, mc.player.getInventory().selectedSlot);
 			EntityUtil.syncInventory();
 		} else if ((pearl = InventoryUtil.findItem(Items.ENDER_PEARL)) != -1) {
 			int old = mc.player.getInventory().selectedSlot;
 			InventoryUtil.switchToSlot(pearl);
-			sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id));
+			sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id, mc.player.getYaw(), mc.player.getPitch()));
 			InventoryUtil.switchToSlot(old);
 		}
 		AutoPearl.throwing = false;
