@@ -38,7 +38,7 @@ public class PearlSpoof extends Module {
 
             mc.getNetworkHandler().sendPacket(new UpdateSelectedSlotC2SPacket(pearlSlot == 1 ? pearlSlot + 1 : pearlSlot - 1));
             mc.getNetworkHandler().sendPacket(new UpdateSelectedSlotC2SPacket(pearlSlot));
-            sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id));
+            sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id, mc.player.getYaw(), mc.player.getPitch()));
             mc.getNetworkHandler().sendPacket(new UpdateSelectedSlotC2SPacket(oldSlot));
         }
         disable();
