@@ -4,8 +4,8 @@ import dev.kizuna.mod.modules.settings.impl.BooleanSetting;
 import dev.kizuna.api.events.eventbus.EventHandler;
 import dev.kizuna.api.events.impl.PlaySoundEvent;
 import dev.kizuna.mod.modules.Module;
-import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
+import net.minecraft.util.Identifier;
 
 import java.util.ArrayList;
 
@@ -22,39 +22,41 @@ public class NoSoundLag extends Module {
 			add(new BooleanSetting("Explode", true));
 	private final BooleanSetting attack =
 			add(new BooleanSetting("Attack", true));
-	static final ArrayList<SoundEvent> armor = new ArrayList<>();
+	static final ArrayList<Identifier> armor = new ArrayList<>();
 	@EventHandler
 	public void onPlaySound(PlaySoundEvent event){
 		if (equip.getValue()) {
-			for (SoundEvent se : armor) {
-				if (event.sound.getId() == se.getId()) {
+			for (Identifier se : armor) {
+				if (event.sound.getId().equals(se)) {
 					event.cancel();
 					return;
 				}
 			}
 		}
 		if (explode.getValue()) {
-			if (event.sound.getId() == SoundEvents.ENTITY_GENERIC_EXPLODE.getId() || event.sound.getId() == SoundEvents.ENTITY_DRAGON_FIREBALL_EXPLODE.getId()) {
+			if (event.sound.getId().equals(Identifier.of("minecraft", "entity.generic.explode"))
+					|| event.sound.getId().equals(Identifier.of("minecraft", "entity.dragon_fireball.explode"))) {
 				event.cancel();
 				return;
 			}
 		}
 		if (attack.getValue()) {
-			if (event.sound.getId() == SoundEvents.ENTITY_PLAYER_ATTACK_WEAK.getId() || event.sound.getId() == SoundEvents.ENTITY_PLAYER_ATTACK_STRONG.getId()) {
+			if (event.sound.getId().equals(Identifier.of("minecraft", "entity.player.attack.weak"))
+					|| event.sound.getId().equals(Identifier.of("minecraft", "entity.player.attack.strong"))) {
 				event.cancel();
             }
 		}
 	}
 
 	static {
-		armor.add(SoundEvents.ITEM_ARMOR_EQUIP_NETHERITE);
-		armor.add(SoundEvents.ITEM_ARMOR_EQUIP_TURTLE);
-		armor.add(SoundEvents.ITEM_ARMOR_EQUIP_CHAIN);
-		armor.add(SoundEvents.ITEM_ARMOR_EQUIP_ELYTRA);
-		armor.add(SoundEvents.ITEM_ARMOR_EQUIP_DIAMOND);
-		armor.add(SoundEvents.ITEM_ARMOR_EQUIP_GOLD);
-		armor.add(SoundEvents.ITEM_ARMOR_EQUIP_IRON);
-		armor.add(SoundEvents.ITEM_ARMOR_EQUIP_LEATHER);
-		armor.add(SoundEvents.ITEM_ARMOR_EQUIP_GENERIC);
+		armor.add(Identifier.of("minecraft", "item.armor.equip_netherite"));
+		armor.add(Identifier.of("minecraft", "item.armor.equip_turtle"));
+		armor.add(Identifier.of("minecraft", "item.armor.equip_chain"));
+		armor.add(Identifier.of("minecraft", "item.armor.equip_elytra"));
+		armor.add(Identifier.of("minecraft", "item.armor.equip_diamond"));
+		armor.add(Identifier.of("minecraft", "item.armor.equip_gold"));
+		armor.add(Identifier.of("minecraft", "item.armor.equip_iron"));
+		armor.add(Identifier.of("minecraft", "item.armor.equip_leather"));
+		armor.add(Identifier.of("minecraft", "item.armor.equip_generic"));
 	}
 }
