@@ -3,7 +3,6 @@ package dev.kizuna.mod.modules.impl.misc;
 import dev.kizuna.api.events.eventbus.EventHandler;
 import dev.kizuna.api.events.impl.PacketEvent;
 import dev.kizuna.mod.modules.Module;
-import dev.kizuna.mod.modules.impl.combat.Criticals;
 import net.minecraft.network.packet.Packet;
 import net.minecraft.network.packet.c2s.play.HandSwingC2SPacket;
 import net.minecraft.network.packet.c2s.play.PlayerInteractEntityC2SPacket;
@@ -18,7 +17,7 @@ public class TrueAttackCooldown extends Module {
     @EventHandler
     public void onPacket(PacketEvent.Send event) {
         Packet<?> packet = event.getPacket();
-        if (packet instanceof HandSwingC2SPacket || packet instanceof PlayerInteractEntityC2SPacket && Criticals.getInteractType((PlayerInteractEntityC2SPacket) packet) == Criticals.InteractType.ATTACK) {
+        if (packet instanceof HandSwingC2SPacket || packet instanceof PlayerInteractEntityC2SPacket) {
             mc.player.resetLastAttackedTicks();
         }
     }
