@@ -11,10 +11,14 @@ import net.minecraft.screen.MerchantScreenHandler;
 import net.minecraft.screen.slot.SlotActionType;
 import net.minecraft.village.TradeOffer;
 import net.minecraft.village.TradeOfferList;
+import net.minecraft.village.TradedItem;
+
+import java.util.Optional;
 
 public class AutoTrade extends Module {
     public AutoTrade() {
         super("AutoTrade", Category.Player);
+        setChinese("自动交易");
     }
     public final SliderSetting repeat = add(new SliderSetting("Repeat", 2, 1, 15, 1));
     public final BooleanSetting autoClose = add(new BooleanSetting("AutoClose", true));
@@ -24,42 +28,45 @@ public class AutoTrade extends Module {
             int i = 0;
             boolean flag = true;
 
-                TradeOfferList list = handler.getRecipes();
-                for (int size = 0; size < list.size(); ++size) {
-                    if (i >= repeat.getValue()) return;
-                    TradeOffer tradeOffer = list.get(size);
-                    if (!tradeOffer.isDisabled()) {
-                        if (Kawaii.TRADE.inWhitelist(tradeOffer.getSellItem().getItem().getTranslationKey())) {
-                            while (i < repeat.getValue() && flag) {
-                                flag = false;
-                                if (!tradeOffer.getAdjustedFirstBuyItem().isEmpty()) {
-                                    int count = InventoryUtil.getItemCount(tradeOffer.getAdjustedFirstBuyItem().getItem());
-                                    if (handler.getSlot(0).getStack().getItem() == tradeOffer.getAdjustedFirstBuyItem().getItem()) {
-                                        count += handler.getSlot(0).getStack().getCount();
-                                    }
-                                    if (count < tradeOffer.getAdjustedFirstBuyItem().getCount()) {
-                                        continue;
-                                    }
+            TradeOfferList list = handler.getRecipes();
+            for (int size = 0; size < list.size(); ++size) {
+                if (i >= repeat.getValue()) return;
+                TradeOffer tradeOffer = list.get(size);
+                if (!tradeOffer.isDisabled()) {
+                    if (Kawaii.TRADE.inWhitelist(tradeOffer.getSellItem().getItem().getTranslationKey())) {
+                        while (i < repeat.getValue() && flag) {
+                            flag = false;
+                            TradedItem firstBuyItem = tradeOffer.getFirstBuyItem();
+                            if (!firstBuyItem.itemStack().isEmpty()) {
+                                int count = InventoryUtil.getItemCount(firstBuyItem.itemStack().getItem());
+                                if (handler.getSlot(0).getStack().getItem() == firstBuyItem.itemStack().getItem()) {
+                                    count += handler.getSlot(0).getStack().getCount();
                                 }
-                                if (!tradeOffer.getSecondBuyItem().isEmpty()) {
-                                    int count = InventoryUtil.getItemCount(tradeOffer.getSecondBuyItem().getItem());
-                                    if (handler.getSlot(1).getStack().getItem() == tradeOffer.getSecondBuyItem().getItem()) {
-                                        count += handler.getSlot(1).getStack().getCount();
-                                    }
-                                    if (count < tradeOffer.getSecondBuyItem().getCount()) {
-                                        continue;
-                                    }
+                                if (count < firstBuyItem.count()) {
+                                    continue;
                                 }
-                                mc.getNetworkHandler().sendPacket(new SelectMerchantTradeC2SPacket(size));
-                                mc.interactionManager.clickSlot(mc.player.currentScreenHandler.syncId, 2, 1, SlotActionType.QUICK_MOVE, mc.player);
-                                flag = true;
-                                i++;
-                                //mc.player.currentScreenHandler.onSlotClick(3, 1, SlotActionType.QUICK_MOVE, mc.player);
                             }
+                            Optional<TradedItem> secondBuyItem = tradeOffer.getSecondBuyItem();
+                            if (secondBuyItem.isPresent()) {
+                                TradedItem secondItem = secondBuyItem.get();
+                                int count = InventoryUtil.getItemCount(secondItem.itemStack().getItem());
+                                if (handler.getSlot(1).getStack().getItem() == secondItem.itemStack().getItem()) {
+                                    count += handler.getSlot(1).getStack().getCount();
+                                }
+                                if (count < secondItem.count()) {
+                                    continue;
+                                }
+                            }
+                            mc.getNetworkHandler().sendPacket(new SelectMerchantTradeC2SPacket(size));
+                            mc.interactionManager.clickSlot(mc.player.currentScreenHandler.syncId, 2, 1, SlotActionType.QUICK_MOVE, mc.player);
+                            flag = true;
+                            i++;
+                            //mc.player.currentScreenHandler.onSlotClick(3, 1, SlotActionType.QUICK_MOVE, mc.player);
                         }
                     }
                 }
-                //CommandManager.sendChatMessage(handler.getSlot(0).getStack().getItem().getName().getString());
+            }
+            //CommandManager.sendChatMessage(handler.getSlot(0).getStack().getItem().getName().getString());
             if (autoClose.getValue() && i < repeat.getValue()) {
                 mc.getNetworkHandler().sendPacket(new CloseHandledScreenC2SPacket(mc.player.currentScreenHandler.syncId));
                 mc.currentScreen.close();
