@@ -4,6 +4,7 @@ import dev.kizuna.api.events.eventbus.EventHandler;
 import dev.kizuna.api.events.impl.PacketEvent;
 import dev.kizuna.mod.modules.Module;
 import dev.kizuna.mod.modules.settings.impl.BooleanSetting;
+import net.minecraft.component.DataComponentTypes;
 import net.minecraft.network.packet.c2s.play.PlayerActionC2SPacket;
 import net.minecraft.network.packet.c2s.play.PlayerInteractItemC2SPacket;
 import net.minecraft.util.Hand;
@@ -18,16 +19,16 @@ public class PacketEat extends Module {
 		INSTANCE = this;
 	}
 
-    @Override
-    public void onUpdate() {
-		if (deSync.getValue() && mc.player.isUsingItem() && mc.player.getActiveItem().getItem().isFood()){
-			Module.sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id));
+	@Override
+	public void onUpdate() {
+		if (deSync.getValue() && mc.player.isUsingItem() && mc.player.getActiveItem().get(DataComponentTypes.FOOD) != null) {
+			Module.sendSequencedPacket(id -> new PlayerInteractItemC2SPacket(Hand.MAIN_HAND, id, mc.player.getYaw(), mc.player.getPitch()));
 		}
 	}
 
 	@EventHandler
 	public void onPacket(PacketEvent.Send event) {
-		if (event.getPacket() instanceof PlayerActionC2SPacket packet && packet.getAction() == PlayerActionC2SPacket.Action.RELEASE_USE_ITEM && mc.player.getActiveItem().getItem().isFood()) {
+		if (event.getPacket() instanceof PlayerActionC2SPacket packet && packet.getAction() == PlayerActionC2SPacket.Action.RELEASE_USE_ITEM && mc.player.getActiveItem().get(DataComponentTypes.FOOD) != null) {
 			event.cancel();
 		}
 	}
